+++ src/components/const/const.js
@@ -1,3 +1,28 @@
+export const FilmRoute = {
+  MAIN: `/`,
+  FILM_INFO: `/film-info`,
+};
+
+export const ActiveMenu = {
+  OVERVIEW: `OVERVIEW`,
+  DETAILS: `DETAILS`,
+  REVIEWS: `REVIEWS`,
+};
+
+export const RatingLevels = {
+  AWESOME: `Awesome`,
+  VERY_GOOD: `Very good`,
+  GOOD: `Good`,
+  NORMAL: `Normal`,
+  BAD: `Bad`,
+};
+
+export const genreType = {
+  ALL: `All genres`,
+};
+
+export const MONTHS = [`January`, `February`, `March`, `April`, `May`, `June`, `July`, `August`, `September`, `October`, `November`, `December`];
+
 export const playerType = {
   MOVIE: `MOVIE`,
   TRAILER: `TRAILER`,
