+++ src/components/film-info-description/film-info-description.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import MovieCardDescriptionNav from "../film-info-description-nav/film-info-nav.jsx";
 import FilmInfoDescriptionOverview from "../film-info-description-overview/film-info-description-overview.jsx";
 import {connect} from "react-redux";
-import {ActiveMenu} from "../utils/utils";
+import {ActiveMenu} from "../const/const.js";
 import FilmInfoDescriptionReviews from "../film-info-description-reviews/film-info-description-reviews.jsx";
 import FilmInfoDescriptionDetails from "../film-info-description-details/film-info-description-details.jsx";
 import PropTypes from "prop-types";
