+++ src/components/film-info/film-info.jsx
@@ -2,9 +2,10 @@ import React from "react";
 import {connect} from "react-redux";
 import PropTypes from "prop-types";
 import history from "../../history/history.js";
-import {filterFilms, FilmRoute} from "../utils/utils.js";
+import {filterFilms} from "../utils/utils.js";
 import FilmInfoDescription from "../film-info-description/film-info-description.jsx";
 import FilmList from "../film-list/film-list.jsx";
+import {FilmRoute} from "../const/const.js";
 
 const FilmInfo = ({activeFilm, films}) => {
   if (JSON.stringify(activeFilm) === `{}`) {
