+++ src/components/utils/utils.js
@@ -1,28 +1,7 @@
+import {MONTHS, genreType, RatingLevels} from "../const/const.js";
+
 export const extend = (a, b) => {
   return Object.assign({}, a, b);
-};
-
-export const FilmRoute = {
-  MAIN: `/`,
-  FILM_INFO: `/film-info`,
-};
-
-export const ActiveMenu = {
-  OVERVIEW: `OVERVIEW`,
-  DETAILS: `DETAILS`,
-  REVIEWS: `REVIEWS`,
-};
-
-export const RatingLevels = {
-  AWESOME: `Awesome`,
-  VERY_GOOD: `Very good`,
-  GOOD: `Good`,
-  NORMAL: `Normal`,
-  BAD: `Bad`,
-};
-
-export const genreType = {
-  ALL: `All genres`,
 };
 
 export const changeFirstLetterUppercase = (word) => word[0].toUpperCase() + word.slice(1);
@@ -60,8 +39,6 @@ export const filterFilms = (data, genres) => {
   }
   return films;
 };
-
-export const MONTHS = [`January`, `February`, `March`, `April`, `May`, `June`, `July`, `August`, `September`, `October`, `November`, `December`];
 
 export const formatDateForReview = (date) => {
   const _date = new Date(date);
